% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allpart.R
\name{allpart}
\alias{allpart}
\title{Concatenate identical properties of different Gadget particle species}
\usage{
allpart(dat, field = "Coordinates", species = seq(0, 5))
}
\arguments{
\item{dat}{list containing the particle data of a single Gadget snapshot, e.g. read via \code{\link{readgadget}}.}

\item{field}{character specifying particle property to concatenate.}

\item{species}{vector of particle species to be included.}
}
\value{
Returns an N-vector or an N-by-3 matrix representing the property "field" of all the particle species given in \code{species} (0-indexed).
}
\description{
Concatenate identical properties of different particle species in a Gadget snapshot
}
\author{
Danail Obreschkow
}
