% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.snapshot.R
\name{as.snapshot}
\alias{as.snapshot}
\title{Create GADGET snapshot object from particle data}
\usage{
as.snapshot(x, v = NULL, m = NULL, id = NULL, types = NULL)
}
\arguments{
\item{x}{list of N-by-3 matrices specifying the coordinates of particles of various types}

\item{v}{optional list of N-by-3 matrices with particle velocities}

\item{m}{optional list of N-vectors with particle masses}

\item{id}{optional list of N-vectors with particle IDs}

\item{types}{optional vector of particle indices. The number of elements in this vector must be the same as the number of matrices in the list \code{x}.}
}
\value{
Returns an object of class \code{snapshot}, which is a structured list that closely resembles the HDF5 format of Gadget (see \url{https://wwwmpa.mpa-garching.mpg.de/gadget4/}).
}
\description{
Converts raw particle data into a \code{snapshot} object.
}
\author{
Danail Obreschkow
}
