% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.gadget.R
\name{compile.gadget}
\alias{compile.gadget}
\title{Compile Gadget simulation}
\usage{
compile.gadget(
  source,
  DIR = NULL,
  CONFIG = NULL,
  EXEC = NULL,
  n.cores = 8,
  clean = FALSE,
  verbose = FALSE,
  measure.time = TRUE
)
}
\arguments{
\item{source}{path of the source code. This path must contain the Makefile ana src directory.}

\item{DIR}{path containing the standard configuration file (\code{Config.sh}) and executable (\code{Gadget4}, or similar) to be written.}

\item{CONFIG}{custom path+filename of configuration file.}

\item{EXEC}{custom path+filename of the executable.}

\item{n.cores}{number of cores to compile the code}

\item{clean}{logical flag indicating whether previous compilations should first be cleaned, if they exist. Make sure to specify the right directory (\code{DIR}).}

\item{verbose}{logical flag indicating whether to display the standard console output}

\item{measure.time}{logical flag indicating whether to measure the total compilation time}
}
\value{
Returns the console command used to compile Gadget.
}
\description{
Complies the Gadget code with custom parameter file.
}
\author{
Danail Obreschkow
}
