% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetry.env.R
\docType{data}
\name{.gadgetry.env}
\alias{.gadgetry.env}
\title{Package environment}
\format{
An object of class \code{environment} of length 1.
}
\usage{
.gadgetry.env
}
\description{
Environment used to store default parameters used in \code{\link{writeparam}}, as well as optional other global variables. The default parameters are stored in \code{.gadgetry.env$default.parameters}.
}
\seealso{
\code{\link{writeparam}}
}
\author{
Danail Obreschkow
}
\keyword{datasets}
