% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halomodel.R
\name{halomodel}
\alias{halomodel}
\title{Standard spherical mass distributions in astrophysics}
\usage{
halomodel(model = "hernquist", rmax = NULL, plots = FALSE, rplot = 10)
}
\arguments{
\item{model}{name of model. Currently only "Hernquist".}

\item{rmax}{truncation radius; profiles are normalized, such that this radius contains the full mass}

\item{plots}{logical flag indicating whether to produce a series of test plots}

\item{rplot}{maximum plotting radius; only used if plots=TRUE.}
}
\value{
list of functions for the selected model
}
\description{
Generates functions associated with spherically symmetric halo models. Currently only Hernquist model implemented. Extension planned. Everything is in units dimensionless units where M=R=G=1, where M is the total halo mass, R is the characteristic radius (or truncation radius if no characteristic radius exists) and G is the gravitational constant.
}
\examples{

f = halomodel(plots = TRUE, rplot = 5, rmax=3)

}
\author{
Danail Obreschkow
}
