% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot4.R
\name{plot4}
\alias{plot4}
\title{Display a simulation snapshot in four different projections}
\usage{
plot4(
  x,
  rotations = c(2, 4, 1, 3),
  screen = TRUE,
  pdffile = NULL,
  title = NULL,
  scale = TRUE,
  width = NULL,
  ...
)
}
\arguments{
\item{x}{Gadget data type, which must contain at least one sublist PartType# (with #=0,1,...). For details, see \code{\link{plot.snapshot}}.}

\item{rotations}{4-vector or 4-element list specifying four different values/vectors of the argument \code{rotation} in \code{\link{plot.snapshot}}. These different rotations are shown in the bottom left, bottom right, top left, and top right panel, respectively, in this order.}

\item{screen}{logical flag specifying whether the images is displayed on the screen.}

\item{pdffile}{optional pdf-filename to save the image as pdf-file.}

\item{title}{Text to be added to the figure.}

\item{scale}{logical flag, specifying if a length scale is shown}

\item{width}{optional horizontal range of the image in the length units of \code{x}. The default corresponds to the full range of particle positions.}

\item{...}{additional parameters for \code{\link{plot.snapshot}}.}
}
\value{
None
}
\description{
Produces a 2-by-2 panel figure, where each panel is made of an \code{\link{plot.snapshot}} image in a different rotation.
}
\examples{
filename = system.file('test_snapshot.hdf5', package='gadgetry')
sn = readsnapshot(filename)
sn$PartType1$smoothing = 10
sn$PartType2$smoothing = 3
sn$PartType2$lum = 0.5
plot4(sn, length.unit='kpc', width=600)

}
\seealso{
\code{\link{plot.snapshot}}
}
