% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.gadget.R
\name{run.gadget}
\alias{run.gadget}
\title{Run Gadget simulation}
\usage{
run.gadget(
  file.gadget,
  file.param,
  n.cores = 8,
  verbose = FALSE,
  measure.time = TRUE
)
}
\arguments{
\item{file.gadget}{path+filename of compiled Gadget-4 code}

\item{file.param}{path+filename of parameter file}

\item{n.cores}{number of cores to be used with MPI}

\item{verbose}{logical flag indicating whether to display the standard console output}

\item{measure.time}{logical flag indicating whether to measure the total simulation time}
}
\value{
Returns the console command used to call Gadget.
}
\description{
Executes a pre-compiled Gadget code with custom parameter file.
}
\author{
Danail Obreschkow
}
