% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeparam.R
\name{writeparam}
\alias{writeparam}
\title{Write Gadget-4 parameter file}
\usage{
writeparam(file, param = NULL)
}
\arguments{
\item{file}{parameter filename}

\item{param}{list of parameter with values}
}
\value{
Returns the full parameter list.
}
\description{
Write parameter file for the astrophysical N-body+SPH simulation code Gadget-4.
}
\author{
Danail Obreschkow
}
