% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writesnapshot.R
\name{writesnapshot}
\alias{writesnapshot}
\title{Write simulation snapshot in Gadget-2 binary format}
\usage{
writesnapshot(part, head, file)
}
\arguments{
\item{part}{is a list of particle properties, for example to be used as initial conditions for a simulation. It contains the items:\cr
\code{x} = N-by-3 matrix specifying the initial position in Cartesian coordinates of the N particles\cr
\code{v} = optional N-by-3 matrix specifying the initial velocities. If not given, velocities are set to zero.\cr
\code{id} = optional N-by-3 matrix specifying unique particle IDs. If not given, the particles will be numbered sequentially from 1 to N.\cr
\code{m} = optional vector giving the masses of the particles, whose mass is not already specified in \code{head$Massarr}. This is only used for variable particles masses. Fixed masses are best specified with the header argument \code{Massarr}.\cr
\code{u} = optional vector giving the internal energies of the SPH particles, i.e. the particles of type 0 (in C-indexing).\cr\cr}

\item{head}{is a list of header arguments. For full details, please refer to the Gadget-2 manual (\url{https://wwwmpa.mpa-garching.mpg.de/gadget/users-guide.pdf}, page 33). The most important items are:\cr
\code{Npart} = 6-vector giving the number of particles of each type. The first element corresponds to gas particles; the other five are treated as collision-less.\cr
\code{NMassarr} = optional 6-vector listing a fixed mass for each particle type. If set to 0 for a type which is present, individual particle masses must be provided in the vector part$m.\cr\cr}

\item{file}{filename.}
}
\value{
Returns a list containing the particle data. The format of the list depends on the input format.
}
\description{
Write astrophysical N-body+SPH snapshot in Gadget-2 binary format, for use as initial conditions with Gadget Version 1/2/3/4 (SnapFormat=1).
}
\seealso{
\code{\link{readsnapshot}}
}
\author{
Danail Obreschkow
}
