% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadgetry-package.R
\docType{package}
\name{gadgetry}
\alias{gadgetry}
\title{Reading, Processing and Visualizing Astrophysical N-Body Simulation Data}
\description{
Routines for post-processing and visualizing astrophysical N-body+SPH simulation data, written to deal with data generated by the Gadget code at <https://wwwmpa.mpa-garching.mpg.de/gadget4/>. The routines are most suitable for smaller simulations (up to ~1e8 particles), as they have been written assuming that the simulation particles of individual snapshots can be loaded into RAM memory at once.
}
\author{
Danail Obreschkow
}
