% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halomodel.R
\name{halomodel}
\alias{halomodel}
\title{Generate standard spherical mass distributions in astrophysics}
\usage{
halomodel(
  model = "hernquist",
  rmax = NULL,
  rscale = 1,
  mtot = 1,
  gravity = 1,
  plots = FALSE,
  rplot = 2
)
}
\arguments{
\item{model}{name of model. Currently available models: "uniform" (spherical top-hat), "isothermal" (singular isothermal sphere with logarithmic potential), "hernquist" (Hernquist 1990), .}

\item{rmax}{optional truncation radius. If not given, a natural model-dependent value is assigned, which is infinite for models where the mass converges in this case.}

\item{rscale}{scale radius for models, which admit a characteristic scale in addition to a possible truncation radius. For all other models, this parameter is ignored. \code{rmax} and \code{rscale} should be provided in the same units.}

\item{mtot}{total mass enclosed by the truncation radius}

\item{gravity}{gravitational constant}

\item{plots}{logical flag indicating whether to produce a series of test plots}

\item{rplot}{maximum plotting radius in the same units as \code{rmax} and \code{rscale}; only used if plots=TRUE.}
}
\value{
list of functions for the selected model
}
\description{
Generates potential-density pairs and auxiliary functions associated with spherically symmetric mass distributions commonly used in astrophysics. The units of all outputs are the same as those of \code{rmax}, \code{rscale}, \code{mtot} and \code{gravity}.
}
\examples{

halomodel('hernquist', plots = TRUE, rplot = 5, rmax=3)

}
\author{
Danail Obreschkow
}
