% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snapshotinterp.R
\name{snapshotinterp}
\alias{snapshotinterp}
\title{Interpolate between Gadget snapshots}
\usage{
snapshotinterp(
  sn0,
  sn1,
  t0 = 0,
  t1 = 1,
  ti = 0.5,
  usevelocities = TRUE,
  v0factor = 1,
  v1factor = NULL,
  afield = NULL,
  dt = NULL
)
}
\arguments{
\item{sn0}{first gadget snapshot}

\item{sn1}{second gadget snapshot}

\item{t0}{time of first snapshot}

\item{t1}{time of second snapshot}

\item{ti}{time of interpolated output snapshot. Values outside the range [t0,t1] lead to an extrapolation.}

\item{usevelocities}{logical flag. If TRUE, velocities are used to improve the interpolation of positions.}

\item{v0factor}{factor to be applied to velocities of sn0 to convert to length/time units of the snapshot coordinates and the t0/t1/ti times.}

\item{v1factor}{factor to be applied to velocities of sn1 to convert to length/time units of the snapshot coordinates and the t0/t1/ti times. If not provided, this will be set equal to \code{v1factor}.}

\item{afield}{optional acceleration field for leapfrog integration. This must be a function of a n-by-3 array representing n position vectors, which returns an n-by-3 array with the n acceleration vectors in the same length and time units as the snapshot coordinates and t0/t1/ti times.}

\item{dt}{optional time step used for leapfrog integration; only used if \code{afield} given. The default is \code{dt=abs(t1-t0)/20}.}
}
\value{
snapshot object with interpolated particle properties.
}
\description{
Interpolate/extrapolate between adjacent Gadget snapshots, building on the interpolation routine \code{\link{particleinterp}}. The ordering of the particles follows that of the nearest snapshot, and if some particles change their species between the two snapshots, the interpolated snapshot uses the species of the nearest snapshot.
}
\author{
Danail Obreschkow
}
